% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceUtils.R
\name{filepath_get_sequence_type}
\alias{filepath_get_sequence_type}
\title{Sequence types of filepaths}
\usage{
filepath_get_sequence_type(filepaths)
}
\arguments{
\item{filepaths}{paths of sequence file (fastq/fasta/other)}
}
\value{
'fastq', 'fasta' or 'Other' depending on the filetype of the paths supplied (character)
}
\description{
Sequence types of filepaths
}
\examples{
filepath_get_sequence_type("file.fastq.gz")
}
