% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_calculate_proportion_of_signal_explained_by_n_strongest_taxids}
\alias{kraken_calculate_proportion_of_signal_explained_by_n_strongest_taxids}
\title{Quantifying Signal Spread Across Taxid}
\usage{
kraken_calculate_proportion_of_signal_explained_by_n_strongest_taxids(
  kraken_report_df,
  parent_taxid,
  n = 1,
  rank = "S"
)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{parent_taxid}{the taxid representing a larger group within which we want to know whether binning of reads is diffuse or focused (integer)}

\item{n}{use the n most binned species/genus'/etc to calculate 'proportion of reads in parent_taxid lineage explained by a focused group' (integer)}

\item{rank}{one of: (S, G, F, O, C, P, K). S=speies, G=genus, etc. Determines which rank in calculation of binning 'diffuseness'. You will almost always want to use species (S), maybe genus.}
}
\value{
Dataframe describing the percentage of reads in parent clade that are explained by the n most frequently classified taxids
}
\description{
This function tells you the proportion of reads belonging to a taxid (or its descendents) whose classification within the larger group is driven by classification into the most commonly classified species/genus/family (user can choose).
Say we notice a high number of bacterial reads and we want to know if this phenomina is driven largely by a single species or a more diffuse binning of reads into lots of different bacterial species.
The low spread binning (most in single species) would increase our confidence that the sample actually has the in it. If binning is more diffuse its more likely to have some other explanation.
}
