% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_confidence_palette}
\alias{kraken_confidence_palette}
\title{Confidence Palette}
\usage{
kraken_confidence_palette()
}
\value{
colours for confidence levels (character)
}
\description{
Confidence Palette
}
\examples{
kraken_confidence_palette()
}
