% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{kraken_database_close_connection}
\alias{kraken_database_close_connection}
\title{Disconnect from Kraken Database}
\usage{
kraken_database_close_connection(kreports_db)
}
\arguments{
\item{kreports_db}{object produced by \code{\link[=kraken_database_connect]{kraken_database_connect()}}}
}
\value{
run for its side effects
}
\description{
Disconnect from Kraken Database
}
\examples{
\dontrun{
# open connection to database
kreports_db = kraken_database_connect("path_to_database")

# do required visualisations/analyses

# close connection when done
kraken_database_close_connection(kreports_db)
}
}
