% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{kraken_database_connect}
\alias{kraken_database_connect}
\title{Connect to Kraken Database}
\usage{
kraken_database_connect(path_to_sqlite_db)
}
\arguments{
\item{path_to_sqlite_db}{path to sqlite db file produced by \code{\link[=kraken_reports_parse_to_sqlite_db]{kraken_reports_parse_to_sqlite_db()}}}
}
\value{
dbplyr object
}
\description{
Connect to Kraken Database
}
