% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenExtractReads.R
\name{kraken_extract_reads}
\alias{kraken_extract_reads}
\title{Extract Reads}
\usage{
kraken_extract_reads(
  kreport_path,
  kraken_stdout_path,
  kraken_input_seqs,
  kraken_input_seqs2 = NULL,
  outfile_prefix,
  taxid,
  include_children,
  threads = 1,
  compress_output_seqfile = TRUE
)
}
\arguments{
\item{kreport_path}{path to kraken report (string)}

\item{kraken_stdout_path}{path to stdout (string)}

\item{kraken_input_seqs}{path to fasta/fastq (optionally compressed) files that were the inputs for the kraken. If input was paired-end sequences, include only paths to forward reads here. Supply any reverse reads to kraken_input_seqs2 (character)}

\item{kraken_input_seqs2}{path to reverse reads. Leave null if input was unpaired (character)}

\item{outfile_prefix}{outfile prefix (string)}

\item{taxid}{Taxonomy ID of reads to extract from input seqs}

\item{include_children}{include children of supplied taxid}

\item{threads}{number of threads to use (string)}
}
\value{
invisibly returns paths to fastq/fasta sequences classifed as \strong{taxid} and (optionally) child taxids if \strong{include_children = TRUE}. (character)
}
\description{
Extract Reads
}
\examples{
\dontrun{
kreport_path = system.file(package="utilitybeltkraken", "testfiles/DRR002014_1.100sequences.minikraken8GB.kreport")
kraken_stdout_path = system.file(package="utilitybeltkraken", "testfiles/DRR002014_1.100sequences.minikraken8GB.stdout.tsv")
kraken_input_seqs = c(
  system.file(package="utilitybeltkraken", "testfiles/DRR002014_1.100sequences.fastq.gz"),
  system.file(package="utilitybeltkraken", "testfiles/DRR002014_2.100sequences.fastq.gz")
)
sequence_outfile_prefix = "mykrakenrun"

kraken_extract_reads(
  kreport_path = kreport_path,
  kraken_input_seqs = kraken_input_seqs,
  outfile_prefix = sequence_outfile_prefix,
  taxid = 562,
  include_children = TRUE,
  kraken_stdout_path = kraken_stdout_path,
  threads = 1
)
}

}
