% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenInspectToTaxonomy.R
\name{kraken_inspect_parse_to_sqlite_db}
\alias{kraken_inspect_parse_to_sqlite_db}
\title{Parse Kraken Inspect File}
\usage{
kraken_inspect_parse_to_sqlite_db(
  file,
  dbname_taxonomy = paste0(getwd(), "/kraken_taxonomic_database.sqlite"),
  dbname_heirarchy = paste0(getwd(), "/kraken_taxonomic_heirarchy.sqlite")
)
}
\arguments{
\item{file}{path to the file produced by running 'kraken2 inspect' on database}
}
\value{
tibble representation of kraken inspect.
}
\description{
Parse Kraken Inspect File
}
\examples{
inspect_file = system.file(package='utilitybeltkraken', "example_data/inspect_pluspf_20210517.txt")
kraken_inspect_parse(inspect_file)
}
