% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_report_add_descendancy_status}
\alias{kraken_report_add_descendancy_status}
\title{Identify Taxid Descendancy Status}
\usage{
kraken_report_add_descendancy_status(
  kraken_report_df,
  taxid,
  columname = NA,
  careful = TRUE,
  verbose = T
)
}
\arguments{
\item{kraken_report_df}{an UNSORTED kraken report dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{taxid}{ncbi taxonomic id (integer)}

\item{columname}{name of the column to add to the dataframe}

\item{careful}{should we do a bunch of quality assertions (slight speed overhead) (flag)}

\item{verbose}{print more informative messages}
}
\value{
This function returns the same dataframe from \strong{kraken_report_df} produced by \link{kraken_reports_parse} or link{kraken_report_parse}  but with  a new columns describing inclusive descendancey status to a particular taxid
}
\description{
Identify taxids in your \strong{kraken_report_df} that are either equal to the user specified taxonomy ID or one of its descendants/children
This is seful for telling what species are bacterial / viral / belong to a particular genus etc.
\strong{WARNING: } this function relies on inherent properties of kraken reports - make sure you run this ONLY on unsorted dataframes produced by \link{kraken_reports_parse} or link{kraken_report_parse}
}
