% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenAnnotateStatistics.R
\name{kraken_report_add_hit_confidence}
\alias{kraken_report_add_hit_confidence}
\title{Stratify Strength of Microbial Hits}
\usage{
kraken_report_add_hit_confidence(kraken_report_df)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}
}
\value{
invisible null (run for its side effects)
}
\description{
Adds a column to the kraken report df that stratifies each microbe-sample hit confidence into a discrete level.
Levels include 'High', 'Medium', 'Low', 'Very Low', 'No Hit'.
Adds column: 'Confidence'
}
