% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenAnnotateStatistics.R
\name{kraken_report_add_median_centered_tau2_scaled_metric}
\alias{kraken_report_add_median_centered_tau2_scaled_metric}
\title{Add MedianTau2 to Report}
\usage{
kraken_report_add_median_centered_tau2_scaled_metric(kraken_report_df)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}
}
\value{
Nothing - input data.table is modified in place
}
\description{
Adds a median centered Tau2 scaled metric to data.table (modifies in place)
}
\seealso{
Other standardised metrics: 
\code{\link{kraken_info_standardised_metrics}()},
\code{\link{kraken_report_add_median_centered_iqr_scaled_metric}()},
\code{\link{kraken_report_add_median_centered_qn_scaled_metric}()},
\code{\link{kraken_report_add_robust_zscore}()},
\code{\link{kraken_report_add_zscore}()}
}
\concept{standardised metrics}
