% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenAnnotateStatistics.R
\name{kraken_report_add_zscore}
\alias{kraken_report_add_zscore}
\title{Add Zscore to Report}
\usage{
kraken_report_add_zscore(kraken_report_df)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}
}
\value{
Nothing - input data.table is modified in place
}
\description{
Add Zscore to Report
}
\seealso{
Other standardised metrics: 
\code{\link{kraken_info_standardised_metrics}()},
\code{\link{kraken_report_add_median_centered_iqr_scaled_metric}()},
\code{\link{kraken_report_add_median_centered_qn_scaled_metric}()},
\code{\link{kraken_report_add_median_centered_tau2_scaled_metric}()},
\code{\link{kraken_report_add_robust_zscore}()}
}
\concept{standardised metrics}
