% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectFromKreportFunctions.R.R
\name{kraken_report_get_child_taxids}
\alias{kraken_report_get_child_taxids}
\title{Get Child Taxids}
\usage{
kraken_report_get_child_taxids(
  path_to_kraken_report,
  taxid,
  inclusive = FALSE,
  exclude_children_without_reads_directly_classified = FALSE
)
}
\arguments{
\item{path_to_kraken_report}{path to kraken report file.}

\item{taxid}{taxonomy ID of species you want to find  the relatives of (int)}

\item{inclusive}{should we include the user-specified taxid in the result (flag)}

\item{exclude_children_without_reads_directly_classified}{in the list of child taxids returned - should we only return taxids that have at least 1 read directly classifed, or return all child taxids irrespective of read counts.}
}
\value{
taxids that the kraken report says are children of the user-supplied taxid (int)
}
\description{
Scan a kraken report to find what taxids kraken considered to be the chldren of others
}
