% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_report_parse}
\alias{kraken_report_parse}
\title{Parse Kraken Reports}
\usage{
kraken_report_parse(path_to_kreport, sample_id = NULL, verbose = TRUE)
}
\arguments{
\item{path_to_kreport}{path to kraken report}

\item{sample_id}{sample identifier. By default will guess from the filename (takes everything before the first '.' as sample name)}
}
\value{
a dataframe describing all samples kraken reports
}
\description{
Parse Kraken Reports
}
