% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_report_parse2}
\alias{kraken_report_parse2}
\title{kraken_report_parse}
\usage{
kraken_report_parse2(kraken2report, sample_id = NA)
}
\arguments{
\item{kraken2report}{path to kraken directory (string)}

\item{sample_id}{(string)}
}
\value{
dataframe describing the kraken report
}
\description{
If you want to import multiple reports use \strong{kraken_report_parses}
}
