% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenTabulate.R
\name{kraken_report_tabulate_hits}
\alias{kraken_report_tabulate_hits}
\title{Title}
\usage{
kraken_report_tabulate_hits(
  kraken_report_df,
  rank = "S",
  zscore_robust_min = 5,
  min_reads = 100
)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{rank}{one of: (S, G, F, O, C, P, K). S=speies, G=genus, etc. Determines which rank in calculation of binning 'diffuseness'. You will almost always want to use species (S), maybe genus.}

\item{zscore_robust_min}{minimum robust zscore to be considered a hit at the lowest confidence level}

\item{min_reads}{minimum reads classified under clade required to be considered a hit at the lowest confidence level}
}
\value{
dataframe describing all hits passing \code{zscore_robust_min} and \code{min_reads} filters, with levels of confidence added (based on robust zscore and the number of reads assigned under clade
}
\description{
Title
}
