% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{kraken_report_to_sqlite_db}
\alias{kraken_report_to_sqlite_db}
\title{Convert Kraken Report Dataframe to sqlite database}
\usage{
kraken_report_to_sqlite_db(
  kraken_report_df,
  database_name = paste0(getwd(), "/kraken_report_database.sqlite"),
  table_name = "kreports",
  overwrite = TRUE
)
}
\arguments{
\item{kraken_report_df}{dataframe from \code{\link[=kraken_reports_parse]{kraken_reports_parse()}}}

\item{database_name}{(If unsure leave defaults) name of sqlite database to create.}

\item{table_name}{(If unsure leave defaults) name of kraken report dataframe table in sqlite database}

\item{overwrite}{overwrite existing database? (boolean)}
}
\value{
Name of the Database Created (string). Primarily Run for its side effects
}
\description{
Convert Kraken Report Dataframe to sqlite database
}
