% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_report_visualise_distributions}
\alias{kraken_report_visualise_distributions}
\title{Visualise Distribution}
\usage{
kraken_report_visualise_distributions(
  kraken_report_df,
  taxids_of_interest,
  metric = "ZscoreRobust",
  show_sample_vlines = FALSE,
  supporting_reads_threshold = 50,
  sample_of_interest = NA,
  use_loggable_version = TRUE,
  seed = 1,
  pointsize = c(1, 2),
  ...
)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{taxids_of_interest}{}
}
\value{

}
\description{
Visualise Distribution
}
