% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_report_visualise_single_sample}
\alias{kraken_report_visualise_single_sample}
\title{Title}
\usage{
kraken_report_visualise_single_sample(
  kraken_report_df,
  samples_of_interest,
  rank = "S",
  taxonomy_id_to_highlight = NULL,
  read_threshold = 50
)
}
\arguments{
\item{read_threshold}{}
}
\value{

}
\description{
Title
}
