% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_report_visualise_single_sample_interactive}
\alias{kraken_report_visualise_single_sample_interactive}
\title{Interactive Single Sample Visualisation}
\usage{
kraken_report_visualise_single_sample_interactive(
  tooltip = c("ScientificName", "TaxonomyID", "ReadsCoveredByClade", "ZscoreRobust",
    "RPM"),
  ...
)
}
\arguments{
\item{tooltip}{aesthetics to include in tooltip c("ScientificName", "TaxonomyID", "ReadsCoveredByClade", "ZscoreRobust", "RPM")}

\item{...}{arguments passed to \strong{kraken_report_visualise_single_sample}}
}
\value{
plotly graph
}
\description{
Interactive Single Sample Visualisation
}
