% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_reports_parse2}
\alias{kraken_reports_parse2}
\title{Parse Kraken Reports}
\usage{
kraken_reports_parse2(kraken2directory)
}
\arguments{
\item{kraken2directory}{path to a directory filled with ONLY kraken2 reports}
}
\value{
a dataframe describing all samples kraken reports
}
\description{
Parse Kraken Reports
}
