% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_write_tsv}
\alias{kraken_write_tsv}
\title{Write Kraken TSV}
\usage{
kraken_write_tsv(
  kraken_report_df,
  filepath = paste0("kraken_report_database_", Sys.Date(), ".tsv")
)
}
\arguments{
\item{kraken_report_df}{an UNSORTED kraken report dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{filepath}{filename/path to write tab-separated dataframe to}
}
\value{
Run for its side effects
}
\description{
Write Kraken TSV
}
