% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenInspectToTaxonomy.R
\name{network_style_to_heirarchy_table}
\alias{network_style_to_heirarchy_table}
\title{Title}
\usage{
network_style_to_heirarchy_table(
  taxonomy_ids,
  parent_taxonomy_ids,
  ranks,
  level,
  sciname
)
}
\arguments{
\item{taxonomy_ids}{taxid}

\item{parent_taxonomy_ids}{taxid parent}

\item{ranks}{taxid rank}

\item{level}{taxid level}
}
\value{
heirarchy table. For each taxid list all their ancestors
}
\description{
Title
}
