% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseFunctions.R
\name{kraken_report_to_sqlite_db}
\alias{kraken_report_to_sqlite_db}
\title{Convert Kraken Report Dataframe to sqlite database}
\usage{
kraken_report_to_sqlite_db(
  kraken_report_df,
  database_name = paste0(getwd(), "/kraken_report_database.sqlite"),
  table_name = "kreports",
  overwrite = TRUE
)
}
\arguments{
\item{kraken_report_df}{an UNSORTED kraken report dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{database_name}{(If unsure leave defaults) name of sqlite database to create.}

\item{table_name}{(If unsure leave defaults) name of kraken report dataframe table in sqlite database}
}
\value{
Name of the Database Created (string). Primarily Run for its side effects
}
\description{
Identify taxids in your \strong{kraken_report_df} that are either equal to the user specified taxonomy ID or one of its descendants/children
This is seful for telling what species are bacterial / viral / belong to a particular genus etc.
\strong{WARNING: } this function relies on inherent properties of kraken reports - make sure you run this ONLY on unsorted dataframes produced by \link{kraken_reports_parse} or link{kraken_report_parse}
}
