% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_report_visualise_distributions}
\alias{kraken_report_visualise_distributions}
\title{Kraken Report Visualise Distributions}
\usage{
kraken_report_visualise_distributions(
  kraken_report_df,
  taxids_of_interest,
  show_y_axis_labels = TRUE,
  show_sample_vlines = FALSE,
  sample_of_interest = NA,
  pointsize = 2,
  pointstroke = 1.5
)
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}}

\item{taxids_of_interest}{taxonomy ids to show as distinct facets (numeric)}

\item{show_y_axis_labels}{show y axis labels? (boolean)}

\item{show_sample_vlines}{show vertical lines indicating positions of samples of interest (boolean)}

\item{sample_of_interest}{which samples are we most interested in - use show_sample_vlines flag to mark these samples (bool)}

\item{...}{other aesthetics passed to aes_string. Mostly used to show more properties in hoverbox once plots are made interactive}
}
\value{
ggplot
}
\description{
Kraken Report Visualise Distributions
}
